Base of documentation
Add some extra configurations and extensions

In particular autodoc which allow you to mention some
module/class/function/ in the docs and will automatically include the
documentation for said object.

I've also activate intersphinx; you may want to set the intersphinx map
for numpy/matplotlib. this allow when you do for example `numpy.array`
to automatically link to numpy documentation of array without having to
provide the URL.

I've set the theme to read the docs sphinx, so locally you will need to
install it; but you don't need to on read_the_docs; it's just so the
appearance is identical when building locally.

I hope that this is enough to build on read the docs; it's often a bit
fiddly to get it properly configured the first time.